# Fixed for non-ascii characters and optimized