This PR removes index access to input strings and instead uses range iteration and runeLen offsets.

It also switches from string representations to rune representations, and replaces toLower conversions with equalFold checks, which are significantly faster.

Performance numbers:
```
benchmark                                          old ns/op     new ns/op     delta
BenchmarkFind/with_unreal_4_(~16K_files)-8         18821066      4926121       -73.83%
BenchmarkFind/with_linux_kernel_(~60K_files)-8     45350582      11973487      -73.60%

benchmark                                          old allocs     new allocs     delta
BenchmarkFind/with_unreal_4_(~16K_files)-8         207238         897            -99.57%
BenchmarkFind/with_linux_kernel_(~60K_files)-8     190717         204            -99.89%

benchmark                                          old bytes     new bytes     delta
BenchmarkFind/with_unreal_4_(~16K_files)-8         5254408       158912        -96.98%
BenchmarkFind/with_linux_kernel_(~60K_files)-8     12121526      38496         -99.68%
```