@ericpauley 2018-04-29
@sahilm Thanks for finding that. I'm ready to hand off development on the unicode feature/tests. Feel free to merge into a feature branch and continue development.