@sahilm 2018-04-29
Unicode handling seems to be A-ok. I haven't been able to find more bugs 😺  Thanks folks!

@lunixbochs I would like to apply some of your perf. optimisations. Would you like to make a pull request or should I go ahead steal your ideas? I'm fine either way :)

P.S. 'Twas awesome to see everyone come together and find/fix bugs. Given enough eyeballs all bugs **truly** are shallow.